% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allign.R
\name{allign}
\alias{allign}
\title{This function loads the reference for a given SangeR object}
\usage{
allign(SangeR)
}
\arguments{
\item{SangeR}{object from get_ref() function.}
}
\value{
Objekt with all information about the alignment.
}
\description{
This function loads the reference for a given SangeR object
}
