% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_ref}
\alias{get_ref}
\title{This function loads the reference for a given SangeR object}
\usage{
get_ref(
  SangeR,
  upstream = 500,
  host = "grch37.ensembl.org",
  dataset = "hsapiens_gene_ensembl",
  biomart = "ensembl"
)
}
\arguments{
\item{SangeR}{object from read.ab1() function.}

\item{upstream}{region before gene to be loaded into ref_seq}

\item{host}{Host to connect to. Defaults to grch37.ensembl.or}

\item{dataset}{Dataset you want to use. To see the different datasets available within a biomaRt you can e.g. do: mart = useMart("ensembl"), followed by listDatasets(mart). Default: hsapiens_gene_ensembl}

\item{biomart}{BioMart database name you want to connect to. Possible database names can be retrieved with the function listMarts. Default: ensembl}
}
\value{
Objekt with all information of the reference.
}
\description{
This function loads the reference for a given SangeR object
}
