% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_routine.R
\name{get_ref_routine}
\alias{get_ref_routine}
\title{This function loads the reference for a given SangeR object from a given mart file}
\arguments{
\item{SangeR}{object from read.ab1() function.}

\item{mart}{mart file from directory}
}
\value{
Objekt with all information of the reference.
}
\description{
This function loads the reference for a given SangeR object from a given mart file
}
