% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygote.R
\name{heterozygote}
\alias{heterozygote}
\title{This function checks for heterozygote Nucleotide and gives back the mutation}
\usage{
heterozygote(hetero, nucleotide)
}
\arguments{
\item{hetero}{heterozygote base R, Y, S, W, K, or M}

\item{nucleotide}{Nucleotide from reference}
}
\value{
Nucleotide for mutation
}
\description{
This function checks for heterozygote Nucleotide and gives back the mutation
}
