% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_offline.R
\name{make_mart}
\alias{make_mart}
\title{This function generates a mart which can be saved and later loaded by get_ref_routine}
\usage{
make_mart(
  genename,
  upstream = 500,
  host = "grch37.ensembl.org",
  dataset = "hsapiens_gene_ensembl",
  biomart = "ensembl"
)
}
\arguments{
\item{genename}{genename of the disired region}

\item{upstream}{Region before gene to be loaded into ref_seq}

\item{host}{Host to connect to. Defaults to grch37.ensembl.or}

\item{dataset}{Dataset you want to use. To see the different datasets available within a biomaRt you can e.g. do: mart = useMart("ensembl"), followed by listDatasets(mart). Default: hsapiens_gene_ensembl}

\item{biomart}{BioMart database name you want to connect to. Possible database names can be retrieved with the function listMarts. Default: ensembl}
}
\value{
Objekt with all information of the reference.
}
\description{
This function generates a mart which can be saved and later loaded by get_ref_routine
}
