% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hist}
\alias{plot_hist}
\title{This function produces plots for all containing tags}
\arguments{
\item{SangeR}{object from allign function.}

\item{POI}{list of points of interesst}
}
\value{
Objekt with plots for the given tags
}
\description{
This function produces plots for all containing tags
}
