% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read.ab1}
\alias{read.ab1}
\title{This function reads in a given abi1-file.}
\usage{
read.ab1(
  filename,
  delimiter = "_",
  ID_pos = 2,
  genename_pos = 1,
  cutoff = 0.05,
  min_seq_len = 20,
  offset = 33
)
}
\arguments{
\item{filename}{location of a ab1 file.}

\item{delimiter}{delimiter in genename/ID of ab1 file}

\item{ID_pos}{position of ID in genename/ID of ab1 file}

\item{genename_pos}{position of genename in genename/ID of ab1 file}

\item{cutoff}{cutoff for Basecall of the fastq. Default: 0.05}

\item{min_seq_len}{minimum sequence length for Basecall of the fastq. Default: 20}

\item{offset}{Offset for Basecall of the fastq. Default: 33}
}
\value{
Objekt with all information genename,bnummer,filename,abif,fastq.
}
\description{
This function reads in a given abi1-file.
}
