% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{This function takes the Aa exchange and reference to give back the Aa mutation}
\usage{
translate(Aa, exchange, reference)
}
\arguments{
\item{Aa}{AminoAcid from the reference}

\item{exchange}{result from heterozygote from the mutation position}

\item{reference}{the reference with the two position before and after the mutation}
}
\value{
Objekt with plots for the given tags
}
\description{
This function takes the Aa exchange and reference to give back the Aa mutation
}
