% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hist}
\alias{plot_hist}
\title{plot_hist module}
\usage{
plot_hist(SangeR)
}
\arguments{
\item{SangeR}{object from allign function.}
}
\value{
Objekt with plots for the given tags
}
\description{
This function produces plots for all containing tags
}
