% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/routine_print.R
\name{plot_hist}
\alias{plot_hist}
\alias{print_plots}
\title{This function produces plots for all containing tags}
\usage{
print_plots(SangeR)
}
\arguments{
\item{SangeR}{object from plot_hist function.}

\item{poi}{file with list of points of interest}
}
\value{
Objekt with plots for the given tags

no return, prints PNGs with histograms in working directory
}
\description{
This function produces plots for all containing tags

This function prints plots for all plots given by plot_hist()
}
